chore: reorganizing cpfcnpj folder (#41)
* chore: reorganizing cpfcnpj folder

* fix: reorganizing tests
