+++ src/cpfCnpj/formatters.ts
@@ -12,9 +12,9 @@ export const formatCnpj = (cnpj: string): string => {
 export const formatCpf = (cpf: string): string => {
   if (typeof cpf !== 'string' || cpf.length !== 11) return ''
 
-  cpf = cpf.replace(/\D/g, '');
-  cpf = cpf.replace(/(\d{3})(\d)/, '$1.$2');
-  cpf = cpf.replace(/(\d{3})(\d)/, '$1.$2');
-  cpf = cpf.replace(/(\d{3})(\d{1,2})$/, '$1-$2');
-  return cpf;
+  cpf = cpf.replace(/\D/g, '')
+  cpf = cpf.replace(/(\d{3})(\d)/, '$1.$2')
+  cpf = cpf.replace(/(\d{3})(\d)/, '$1.$2')
+  cpf = cpf.replace(/(\d{3})(\d{1,2})$/, '$1-$2')
+  return cpf
 }
